demo/requirements.txt
pandas
sqlalchemy
python-dotenv
streamlit
plotly
psycopg2-binary
